requirements.txt
numpy==1.17.0
opencv_contrib_python==3.4.2.17
torch==1.4.0
tqdm==4.34.0
matplotlib==3.0.1
scikit_image==0.15.0
scipy==1.1.0
PyYAML==5.3
scikit_learn==0.22.1
skimage==0.0
